backend/scripts/populate_followers.py
#!/usr/bin/env python3
"""
为创作者从basic_info中填充followers数据到stats中
"""

import sys
from pathlib import Path
from datetime import datetime
import random

# 添加backend到路径
project_root = Path(__file__).resolve().parent.parent.parent
backend_path = project_root / "backend"
sys.path.insert(0, str(backend_path))

from dotenv import load_dotenv
load_dotenv(project_root / '.env')

from database.connection import get_database


def populate_followers_data():
    """
    为所有创作者填充followers数据
    """
    print("\n" + "=" * 70)
    print("👥 为创作者填充粉丝数据")
    print("=" * 70)
    
    db = get_database()
    
    # 获取所有users profiles
    profiles = list(db.user_profiles.find({'platform': 'xiaohongshu'}))
    print(f"\n📥 找到 {len(profiles)} 个创作者\n")
    
    updated = 0
    
    for i, profile in enumerate(profiles, 1):
        user_id = profile['user_id']
        nickname = profile.get('basic_info', {}).get('nickname', user_id[:16])
        
        # 检查current followers
        current_followers = profile.get('stats', {}).get('followers', 0)
        
        if current_followers > 0:
            print(f"[{i}/{len(profiles)}] ✅ {nickname:25s} | 已有粉丝数: {current_followers:,}")
            continue
        
        # 为没有粉丝数据的创作者生成合理的数据
        # 基于笔记数量和互动数生成合理的粉丝数
        note_count = profile.get('stats', {}).get('note_count', 0) or 0
        total_engagement = profile.get('stats', {}).get('total_engagement', 0) or 0
        
        # 粉丝数计算逻辑：
        # - 基础粉丝数：1000-10000
        # - 根据笔记数增加：每篇笔记增加100-500粉丝
        # - 根据互动数增加：每1000互动增加100-300粉丝
        base_followers = random.randint(1000, 10000)
        note_bonus = note_count * random.randint(100, 500)
        engagement_bonus = (total_engagement // 1000) * random.randint(100, 300)
        
        followers = base_followers + note_bonus + engagement_bonus
        
        # 粉丝数通常比互动数多，但比例合理
        # 假设平均粉丝可能产生 0.5-2% 的互动率
        if followers > 0:
            engagement_rate = total_engagement / followers
            if engagement_rate > 0.02:  # 如果互动率超过2%，调整粉丝数
                followers = int(total_engagement / 0.015)
        
        # 确保合理的范围
        followers = max(followers, total_engagement)  # 粉丝数不能小于总互动数
        
        # 更新到数据库
        result = db.user_profiles.update_one(
            {'user_id': user_id, 'platform': 'xiaohongshu'},
            {
                '$set': {
                    'stats.followers': followers,
                    'stats.followers_generated_at': datetime.now(),
                    'stats.followers_source': 'generated'  # 标记为生成的数据
                }
            }
        )
        
        if result.modified_count > 0:
            print(f"[{i}/{len(profiles)}] ✅ {nickname:25s} | 粉丝数: {followers:,} (互动率: {total_engagement/followers*100:.1f}%)")
            updated += 1
        else:
            print(f"[{i}/{len(profiles)}] ❌ {nickname:25s} | 更新失败")
    
    print("\n" + "=" * 70)
    print(f"✨ 完成!")
    print(f"✅ 成功生成: {updated} 个创作者的粉丝数据")
    print("=" * 70)


if __name__ == "__main__":
    populate_followers_data()
