backend/requirements.txt
# Web框架（FastAPI核心）
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/DeepSeek API
openai>=1.30.0

# 本地Embeddings模型（⚠️ 不使用DeepSeek API）
sentence-transformers>=2.2.2
torch>=2.0.0
FlagEmbedding>=1.2.0

# 数据库
pymongo==4.6.1
dnspython==2.5.0

# 环境变量与配置管理
python-dotenv==1.0.0
pydantic-settings==2.1.0

# 定时任务调度
APScheduler==3.10.4

# 工具库
requests==2.31.0
duckduckgo-search
